feat(project): move types in common package
