+++ packages/common/types/access.ts
@@ -1,0 +1,4 @@
+export type AccessTokensResponse = {
+  passport: string;
+  refresh_token: string;
+};
