+++ packages/common/types/errors.ts
@@ -1,0 +1,8 @@
+export type JWError = {
+  code: string;
+  description: string;
+};
+
+export type JWErrorResponse = {
+  errors: JWError[];
+};
