+++ platforms/access-bridge/package.json
@@ -21,6 +21,7 @@
   },
   "dependencies": {
     "dotenv": "^16.4.5",
-    "jsonwebtoken": "^9.0.2"
+    "jsonwebtoken": "^9.0.2",
+    "@jwp/ott-common": "*"
   }
 }
