+++ platforms/access-bridge/src/services/AccessService.ts
@@ -1,15 +1,10 @@
 import jwt from 'jsonwebtoken';
+import { AccessTokensResponse } from '@jwp/ott-common/types/access.js';
+import { AccessControlPlan } from '@jwp/ott-common/types/plans.js';
 
 import { ACCESS_CONTROL_CLIENT, API_SECRET } from '../appConfig.js';
 import { BadRequestError, ForbiddenError, NotFoundError, ParameterInvalidError, isJWError } from '../errors.js';
 import { put } from '../http.js';
-
-import { AccessControlPlan } from './PlansService.js';
-
-type AccessTokensResponse = {
-  passport: string;
-  refresh_token: string;
-};
 
 /**
  * AccessService handles interactions with the access management APIs.
